apps/backend/requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
asyncpg==0.31.0
neo4j==6.0.3
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.5.0
google-generativeai==0.8.3
PyJWT==2.10.1
